restructure ray syncer
