+++ src/ray/common/BUILD
@@ -248,10 +248,18 @@ ray_cc_library(
     name = "ray_syncer",
     srcs = [
         "ray_syncer/ray_syncer.cc",
+        "ray_syncer/node_state.cc",
+        "ray_syncer/ray_syncer_client.cc",
+        "ray_syncer/ray_syncer_server.cc",
     ],
     hdrs = [
         "ray_syncer/ray_syncer.h",
-        "ray_syncer/ray_syncer-inl.h",
+        "ray_syncer/node_state.h",
+        "ray_syncer/common.h",
+        "ray_syncer/ray_syncer_bidi_reactor.h",
+        "ray_syncer/ray_syncer_bidi_reactor_base.h",
+        "ray_syncer/ray_syncer_client.h",
+        "ray_syncer/ray_syncer_server.h",
     ],
     deps = [
         ":asio",
